Code cleanup, expansion and ready to evaluation

+++ src/run_evaluation.py
@@ -1,5 +1,6 @@
 import ast
 import json
+import os
 import re
 import time
 import pandas as pd
@@ -7,17 +8,20 @@ import pandas as pd
 from data_service_bird.database import GetDataFromDatabase
 from result_averager import average_results
 
+from data_service_generator import DataServiceGenerator
 from main import LLMAgent
 from evaluation.match_similarity import match_similarity
 
-def run_evaluation_ground_truth(database, model, automatic):
-    llm = LLMAgent(model=model, mode="check_ground_truth", automatic=automatic, database=database)
+def get_queries(database):
+    all_queries = json.load(open(f"queries/train.json"))
+    return [query for query in all_queries if query["db_id"] == database]
+
+def run_evaluation_ground_truth(database, enterprise, model, queries, automatic):
+    llm = LLMAgent(enterprise=enterprise, model=model, mode="check_ground_truth", automatic=automatic, database=database)
     llm_chain = llm.get_chain_truth()
 
-    main_queries = json.load(open(f"queries/test/{database}.json"))
-
     res_eval = []
-    for index, query in enumerate(main_queries):
+    for index, query in enumerate(queries):
         sql = query["SQL"]
     
         input_file = {
@@ -30,11 +34,11 @@ def run_evaluation_ground_truth(database, model, automatic):
         res = llm_chain.invoke(input_file)
         res_eval.append([index, res["output"], res["ground_truth"]])
     res_df = pd.DataFrame(res_eval, columns=["index", "tools_prediction", "ground_truth"])
-    res_df.to_csv(f"evaluation/evaluation_results_check_ground_truth_{database}_{model}.csv", sep=',', index=False)
+    res_df.to_csv(f"evaluation/evaluation_results_check_ground_truth_{database}_{enterprise}_{model}.csv", sep=',', index=False)
     
-def evaluate_ground_truth(database, model):
+def evaluate_ground_truth(database, enterprise, model):
     metrics_res = []
-    eval_results = pd.read_csv(f"evaluation/evaluation_results_check_ground_truth_{database}_{model}.csv")
+    eval_results = pd.read_csv(f"evaluation/evaluation_results_check_ground_truth_{database}_{enterprise}_{model}.csv")
 
     for index, line in eval_results.iterrows():
         ground_truth = ast.literal_eval(line["ground_truth"])
@@ -55,23 +59,21 @@ def evaluate_ground_truth(database, model):
         metrics_res.append([index, accuracy, recall])
     
     metrics_res = pd.DataFrame(metrics_res, columns=["index", "accuracy", "recall"])
-    averages = average_results(metrics_res)
-    metrics_res.to_csv(f"evaluation/detailed_results_check_ground_truth_{database}_{model}.csv", sep=',', index=False)
-    averages.to_csv(f"evaluation/summarized_results_check_ground_truth_{database}_{model}.csv", sep=',', index=False)
+    averages = average_results(metrics_res, "ground_truth_check")
+    metrics_res.to_csv(f"evaluation/detailed_results_check_ground_truth_{database}_{enterprise}_{model}.csv", sep=',', index=False)
+    averages.to_csv(f"evaluation/summarized_results_check_ground_truth_{database}_{enterprise}_{model}.csv", sep=',', index=False)
     print("Detailed metrics are:")
     print(metrics_res)
     print("Summarized metrics are:")
     print(averages)
         
 
-def run_evaluation(database, model, mode, second_mode, services_mode = None, automatic=False):
-    llm = LLMAgent(model, mode, second_mode, services_mode = services_mode, automatic=automatic, database=database)
+def run_evaluation(database, queries, enterprise, model, pipeline_mode, evidence_mode, dataservice_mode = None, automatic=False):
+    llm = LLMAgent(enterprise, model, pipeline_mode, evidence_mode, services_mode = dataservice_mode, automatic=automatic, database=database)
     llm_chain = llm.get_chain()
 
-    main_queries = json.load(open(f"queries/test/{database}.json"))
-
     res_eval = []
-    for index, query in enumerate(main_queries):
+    for index, query in enumerate(queries):
         
         sql = query["SQL"]
     
@@ -98,11 +100,10 @@ def run_evaluation(database, model, mode, second_mode, services_mode = None, aut
         res_eval.append(res_elem)
 
     res_df = pd.DataFrame(res_eval, columns=["index", "question", "sql", "data_services", "pipeline", "output", "output_json"])
-    res_df.to_csv(f"evaluation/evaluation_results_{database}_{model}_{mode}.csv", sep=',', index=False)
+    res_df.to_csv(f"evaluation/evaluation_results_{database}_{enterprise}_{model}_{pipeline_mode}_{evidence_mode}_{dataservice_mode}.csv", sep=',', index=False)
 
 
-def evaluate_results(database, model, mode, automatic):
-    queries = json.load(open(f"queries/test/{database}.json"))
+def evaluate_results(database, queries, enterprise, model, pipeline_mode, evidence_mode, dataservice_mode, automatic):
     db = GetDataFromDatabase()
     
     if automatic:
@@ -111,7 +112,7 @@ def evaluate_results(database, model, mode, automatic):
         db.open_connection(f"data_service_bird/{database}/{database}.sqlite")
 
     metrics_res = []
-    eval_results = pd.read_csv(f"evaluation/evaluation_results_{database}_{model}_{mode}.csv")
+    eval_results = pd.read_csv(f"evaluation/evaluation_results_{database}_{enterprise}_{model}_{pipeline_mode}_{evidence_mode}_{dataservice_mode}.csv")
     #eval_results = pd.read_csv(f"evaluation/evaluation_results_{mode}.csv")
     for index, query in enumerate(queries):
         metrics_res_q_idx = []
@@ -167,9 +168,9 @@ def evaluate_results(database, model, mode, automatic):
         metrics_res.append(metrics_res_q_idx)
 
     metrics_res = pd.DataFrame(metrics_res, columns=["index", "precision", "recall", "acc_cell", "acc_row"])
-    averages = average_results(metrics_res)
-    metrics_res.to_csv(f"evaluation/metrics_results_{database}_{model}_{mode}.csv", sep=',', index=False)
-    averages.to_csv(f"evaluation/summarized_results_{database}_{model}_{mode}.csv", sep=',', index=False)
+    averages = average_results(metrics_res, "pipelines")
+    metrics_res.to_csv(f"evaluation/metrics_results_{database}_{enterprise}_{model}_{pipeline_mode}_{evidence_mode}_{dataservice_mode}.csv", sep=',', index=False)
+    averages.to_csv(f"evaluation/summarized_results_{database}_{enterprise}_{model}_{pipeline_mode}_{evidence_mode}_{dataservice_mode}.csv", sep=',', index=False)
     print("Detailed metrics are:")
     print(metrics_res)
     print("Summarized metrics are:")
@@ -178,24 +179,44 @@ def evaluate_results(database, model, mode, automatic):
 
 if __name__ == "__main__":
     
-    model="Mistral"
-    print(f"Model: {model}")
-    databases=["human_resources"]
-    #database="european_football_1"
-    automatic = False
+    enterprise="Mistral"
+    model = "mistral-large-latest"
+    database="chicago_crime"
+    print(f"Model: {model}, Database: {database}")
     
-    with open("result.json", "w") as file:
-    # Use the `truncate()` method to clear the file's content
-        file.truncate()
+    generation = True
+    force_generation = True
+    ground_truth_check = True
+    pipeline_check = True
+    automatic = True
     
-    #run_evaluation_ground_truth(database, model, automatic=automatic)
-    #evaluate_ground_truth(database, model)
+    pipeline_mode = "wo_pipeline_view"
+    evidence_mode = "added_evidence"
+    dataservice_mode = "ground_truth"
     
-    modes = ["wo_pipeline_view"]
-    second_mode = "added_evidence"
-    for mode in modes:
-        for database in databases:
-            run_evaluation(database, model, mode, second_mode, services_mode="ground_truth", automatic=automatic)
-            evaluate_results(database, model, mode, automatic=automatic)
-        pass
+    queries = get_queries(database)
+    print("Got queries")
+    
+    if generation:
+        exist = os.path.exists(f"data_service_bird_automatic/train_databases/{database}/data_services/{enterprise}/{model}/")
+        if not exist or (exist and force_generation):
+            print("Generating Data services")
+            dataservice_generator = DataServiceGenerator(enterprise, model)
+            dataservice_generator.create_data_services(database)
+
+    if ground_truth_check:
+        print("Performing ground truth check")
+        run_evaluation_ground_truth(database, enterprise, model, queries, automatic=automatic)
+        evaluate_ground_truth(database, enterprise, model)
+        
+    if pipeline_check:
+        print("Performing pipeline check")
+    
+        with open("result.json", "w") as file:
+        # Use the `truncate()` method to clear the file's content
+            file.truncate()
+        
+        run_evaluation(database, queries, enterprise, model, pipeline_mode, evidence_mode, dataservice_mode=dataservice_mode, automatic=automatic)
+        evaluate_results(database, queries, enterprise, model, pipeline_mode, evidence_mode, dataservice_mode, automatic=automatic)
+
